Chapter 7: Function Declarations
In our app, we're going to need to be able to create some zombies. Let's create a function for that.

Create a public function named createZombie. It should take two parameters: _name (a string), and _dna (a uint). Don't forget to pass the first argument by value by using the memory keyword