Chapter 10: More on Functions
We're going to want a helper function that generates a random DNA number from a string.

Create a private function called _generateRandomDna. It will take one parameter named _str (a string), and return a uint. Don't forget to set the data location of the _str parameter to memory.

This function will view some of our contract's variables but not modify them, so mark it as view.